% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_table_components.R
\name{assemble_table_components}
\alias{assemble_table_components}
\title{assemble_table_components}
\usage{
assemble_table_components(table_components)
}
\arguments{
\item{table_components}{produced by \code{process_sheet}}
}
\description{
Reshapes the table components to create a tidy data frame.
}
\examples{

 \donttest{tidyABS_example("australian-industry.xlsx") \%>\% process_sheet(sheets = "Table_1") \%>\% assemble_table_components()  }



}
