% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cell_ref_2_df}
\alias{cell_ref_2_df}
\title{Convert an adress to row/col tibble}
\usage{
cell_ref_2_df(cell_ref)
}
\arguments{
\item{cell_ref}{Cell reference in address format - e.g. "A1"}
}
\description{
Convert an adress reference to a tibble identifying row and col
}
