% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_direction.R
\name{change_direction}
\alias{change_direction}
\title{Change direction
The function changes the direction associated with a col_group/row_group in a processed_sheet - the output of \code{process_sheet}.}
\usage{
change_direction(processed_sheet, group, new_direction)
}
\arguments{
\item{processed_sheet}{returned by process_sheet}

\item{group}{row_group name / col_group name}

\item{new_direction}{unpivotr compass direction}

\item{col_groups}{format object read in by \code{tidyxl::xlsx_cells}}

\item{added_row_groups}{format object read in by \code{tidyxl::xlsx_cells}}
}
\description{
Change direction
The function changes the direction associated with a col_group/row_group in a processed_sheet - the output of \code{process_sheet}.
}
