% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_tidytable}
\alias{create_tidytable}
\title{create_tidytable}
\usage{
create_tidytable(header_df, rowname_df, meta_df, tabledata)
}
\description{
Reshapes the data using unpivotr functions (which are specified in the head dataframes)
}
