% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhead_tabledata.R
\name{enhead_tabledata}
\alias{enhead_tabledata}
\title{Enhead tabledata}
\usage{
enhead_tabledata(header_data, direction, values = tabledata)
}
\arguments{
\item{header_data}{path to .xlsx file}

\item{direction}{sheet nominated for tidying}

\item{values}{sheet nominated for tidying}
}
\description{
Reshapes the data using unpivotr functions (which are specified in the head dataframes)
}
