% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cells_format_features.R
\name{get_indent}
\alias{get_indent}
\title{Get indent of cells}
\usage{
get_indent(local_format_id, formats)
}
\arguments{
\item{local_format_id}{the local_format_id provided by \code{tidyxl::xlsx_cells}}

\item{formats}{sheet object read in by \code{tidyxl::xlsx_formats}}
}
\description{
Return the level of indenting of a specified cell.
}
