% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_table_references}
\alias{get_table_references}
\title{Get table references}
\usage{
get_table_references(x)
}
\arguments{
\item{x}{sheet object read in by \code{tidyxl::xlsx_cells}}
}
\description{
Identify which cells are the numeric table cells by finding the corners
}
