% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_tabledata}
\alias{get_tabledata}
\title{get_tabledata}
\usage{
get_tabledata(sheet, table_ref)
}
\arguments{
\item{sheet}{sheet object read in by \code{tidyxl::xlsx_cells}}

\item{table_ref}{data frame representing corners of numeric cells in excel sheet}
}
\description{
Extracts the numeric data from the table.
}
