% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_tidy_table}
\alias{get_tidy_table}
\title{get tidy table}
\usage{
get_tidy_table(path, sheets)
}
\arguments{
\item{path}{path to .xlsx file}

\item{sheets}{sheet nominated for tidying}
}
\description{
Reshapes the data using unpivotr functions (which are specified in the head dataframes)
}
