% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_tidyABS_components}
\alias{get_tidyABS_components}
\title{get tidyABS components}
\usage{
get_tidyABS_components(path, sheets)
}
\arguments{
\item{path}{path to .xlsx file}

\item{sheets}{sheet nominated for tidying}
}
\description{
Produces the various tidyABS compentents
}
