% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_value_references}
\alias{get_value_references}
\title{Get table references}
\usage{
get_value_references(sheet)
}
\arguments{
\item{sheet}{sheet object read in by \code{tidyxl::xlsx_cells}}
}
\description{
Identify which cells are the numeric table cells by finding the corners
}
