% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{inspect_table_components}
\alias{inspect_table_components}
\title{Inspect table components}
\usage{
inspect_table_components(abs_sheet_processed)
}
\arguments{
\item{abs_sheet_processed}{path to .xlsx file}
}
\description{
Produces the various tidyABS compentents
}
