% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recontruct_table.R
\name{reconstruct_table}
\alias{reconstruct_table}
\title{reconstruct_table}
\usage{
reconstruct_table(table_components)
}
\arguments{
\item{table_components}{object returned by \code{process_ABS_sheet}}
}
\description{
Reconstructs the table components to structure of oringinal excel file.
}
