% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tidyABS_example}
\alias{tidyABS_example}
\title{Get path to tidyABS example
tidyABS comes bundled with some example files in its \code{inst/extdata}
directory. This function make them easy to access.
This function has been copied from the readxl package.}
\usage{
tidyABS_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
Get path to tidyABS example
tidyABS comes bundled with some example files in its \code{inst/extdata}
directory. This function make them easy to access.
This function has been copied from the readxl package.
}
\examples{
readxl_example()
readxl_example("datasets.xlsx")
}
