% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_table_components.R
\name{assemble_table_components}
\alias{assemble_table_components}
\title{assemble_table_components}
\usage{
assemble_table_components(table_componsents)
}
\arguments{
\item{table_componsents}{}
}
\description{
Reshapes the data using unpivotr functions (which are specified in the head dataframes)
}
