% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_direction.R
\name{change_direction}
\alias{change_direction}
\title{Change direction
Changes the direction associated with a col_group/row_group in a \code{process_ABS_sheet}}
\usage{
change_direction(processed_ABS_sheet, group, new_direction)
}
\arguments{
\item{processed_ABS_sheet}{returned by process_ABS_sheet}

\item{group}{row_group name / col_group name}

\item{new_direction}{unpivotr compass direction}

\item{col_groups}{format object read in by \code{tidyxl::xlsx_cells}}

\item{added_row_groups}{format object read in by \code{tidyxl::xlsx_cells}}
}
\description{
Change direction
Changes the direction associated with a col_group/row_group in a \code{process_ABS_sheet}
}
