% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_low_col_names.R
\name{check_low_col_names}
\alias{check_low_col_names}
\title{check_low_col_names}
\usage{
check_low_col_names(col_groups)
}
\arguments{
\item{table_componsents}{}
}
\description{
Checks whether merging of a row label is leading to dupplicates in col and row labels.
}
