% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tidytable.R
\name{create_tidytable}
\alias{create_tidytable}
\title{create_tidytable}
\usage{
create_tidytable(col_groups, row_groups, meta_df, tabledata)
}
\arguments{
\item{tabledata}{}
}
\description{
Reshapes the data using unpivotr functions (which are specified in the head dataframes)
}
