% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tidytable.R
\name{create_tidytable}
\alias{create_tidytable}
\title{create_tidytable}
\usage{
create_tidytable(col_groups, row_groups, meta_df, tabledata)
}
\arguments{
\item{col_groups}{output of \code{process_sheet}}

\item{row_groups}{output of \code{process_sheet}}

\item{meta_df}{output of \code{process_sheet}}

\item{tabledata}{output of \code{process_sheet}}
}
\description{
This is a utility function that reshapes the data using unpivotr functions (which are specified in the head dataframes)
}
