% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhead_tabledata.R
\name{enhead_tabledata}
\alias{enhead_tabledata}
\title{Enhead tabledata}
\usage{
enhead_tabledata(header_data, direction, values = tabledata)
}
\arguments{
\item{header_data}{cells representing headers}

\item{direction}{orietations of heades with respect to value cells.}

\item{values}{centre components of the table, representing values}
}
\description{
This is a utility function that reshapes the data using unpivotr functions (which are specified in the head dataframes)
}
