% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_table_components.R
\name{inspect_table_components}
\alias{inspect_table_components}
\title{Inspect table components}
\usage{
inspect_table_components(processed_sheet)
}
\arguments{
\item{process_sheet}{path to .xlsx file}
}
\description{
This function produces the various tidyABS components.
}
\examples{

 \donttest{tidyABS_example("australian-industry.xlsx") \%>\% process_sheet(sheets = "Table_1") \%>\% inspect_table_components() }

}
