% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_table_components}
\alias{plot_table_components}
\title{Plot table components}
\usage{
plot_table_components(abs_sheet_processed)
}
\arguments{
\item{abs_sheet_processed}{path to .xlsx file}
}
\description{
Produces the various tidyABS compentents
}
