% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ABS_sheet.R
\name{process_sheet}
\alias{process_sheet}
\title{get tidyABS components}
\usage{
process_sheet(path, sheets, manual_value_references = NULL,
  added_row_groups = NULL, keep_meta_data = FALSE)
}
\arguments{
\item{path}{path to .xlsx file}

\item{sheets}{sheet nominated for tidying}
}
\description{
Produces the various tidyABS table components
}
