% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_table.R
\name{reconstruct_table}
\alias{reconstruct_table}
\title{reconstruct_table}
\usage{
reconstruct_table(table_components)
}
\arguments{
\item{table_components}{object returned by \code{process_sheet}}
}
\description{
This function reconstructs the table components from \code{process_sheet} to structure of oringinal excel file.
}
\examples{

 \donttest{tidyABS_example("australian-industry.xlsx") \%>\% process_sheet(sheets = "Table_1") \%>\% reconstruct_table() }





}
