% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_ABS_sheet.R
\name{tidy_ABS_sheet}
\alias{tidy_ABS_sheet}
\title{get tidy table}
\usage{
tidy_ABS_sheet(path, sheets, manual_value_references = NULL)
}
\arguments{
\item{path}{path to .xlsx file}

\item{sheets}{sheet nominated for tidying}
}
\description{
Reshapes the data using unpivotr functions (which are specified in the head dataframes)
}
